Graph-db/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mortgage-graph-platform"
version = "0.1.0"
description = "Production-grade Neo4j mortgage knowledge graph platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "neo4j>=5.28.0",
  "psycopg[binary]>=3.2.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "pandas>=2.3.0",
  "pyarrow>=21.0.0",
  "rdflib>=7.2.0",
  "PyYAML>=6.0.2",
  "tenacity>=9.1.2",
  "streamlit>=1.43.0"
]

[project.optional-dependencies]
test = [
  "pytest>=8.4.0",
  "httpx>=0.28.1"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
